+++ paper/paper.bib
@@ -15,6 +15,7 @@
   pages={1--9},
   year={2016},
   publisher={Nature Publishing Group}
+  doi = {10.1038/sdata.2016.18},
 }
 
 @article{Gregory2018,
@@ -30,7 +31,6 @@
     pages = {1-7},
     abstract = {null},
     number = {4},
-
 }
 
 @article{Contaxis2022,
@@ -41,7 +41,8 @@
   number={2},
   pages={e1009768},
   year={2022},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi = {10.1371/journal.pcbi.1009768},
 }
 
 @article{bauermeister2023evaluating,
@@ -53,10 +54,11 @@
   pages={605--615},
   year={2023},
   publisher={Springer}
+  doi = {10.1007/s10654-023-00997-3},
 }
 
 @article{bauermeister2023ready,
-  title={ready data: the C-Surv data model},
+  title={Research-ready data: the C-Surv data model},
   author={Bauermeister, Sarah and Bauermeister, Joshua R and Bridgman, Ruth and Felici, Caterina and Newbury, Mark and North, Laura and Orton, Christopher and Squires, Emma and Thompson, Simon and Young, Simon and others},
   journal={European Journal of Epidemiology},
   volume={38},
@@ -64,6 +66,7 @@
   pages={179--187},
   year={2023},
   publisher={Springer}
+  doi = {10.1007/s10654-022-00916-y},
 }
 
 
@@ -75,6 +78,7 @@
   pages={601--611},
   year={2020},
   publisher={Springer}
+  doi = {10.1007/s10654-020-00633-4},
 }
 
 @misc{ADDI, 
@@ -93,199 +97,222 @@
  access={24/02/2024}
 } 
 
-@Manual{R-Base,
+@misc{R-Base,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
-  year = {2023},
+  year = {2024},
   url = {https://www.R-project.org/},
+  note = {No DOI identified},
 }
 
-@Manual{R-base,
+@misc{R-base,
   title = {R: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
-  year = {2023},
+  year = {2024},
   url = {https://www.R-project.org/},
+  note = {No DOI identified},
 }
 
-@Manual{R-dplyr,
+@misc{R-dplyr,
   title = {dplyr: A Grammar of Data Manipulation},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller and Davis Vaughan},
   year = {2023},
   note = {R package version 1.1.4},
   url = {https://dplyr.tidyverse.org},
+  note = {No DOI identified},
 }
 
-@Manual{R-DT,
+@misc{R-DT,
   title = {DT: A Wrapper of the JavaScript Library DataTables},
   author = {Yihui Xie and Joe Cheng and Xianying Tan},
   year = {2024},
   note = {R package version 0.32},
   url = {https://github.com/rstudio/DT},
+  note = {No DOI identified},
 }
 
-@Manual{R-fontawesome,
+@misc{R-fontawesome,
   title = {fontawesome: Easily Work with Font Awesome Icons},
   author = {Richard Iannone},
   year = {2023},
   note = {R package version 0.5.2, 
 https://rstudio.github.io/fontawesome/},
   url = {https://github.com/rstudio/fontawesome},
[…]
