### 2024-04-22 #6454 @RoryChenXY
> @editorialbot check references
>
> 👋 @RoryChenXY - while we are waiting for reviewers to sign on, could you please check/fix the missing DOI's listed below from the citations in your paper? Thanks!

Dear Chris @crvernon ,
I have updated the .bib file in my repository the best I can, however, the websites and many R packages do not have DOI.
Apologies for the inconvenience.
