+++ paper/paper.bib
@@ -14,7 +14,7 @@
   number={1},
   pages={1--9},
   year={2016},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi = {10.1038/sdata.2016.18},
 }
 
