+++ paper/paper.bib
@@ -53,7 +53,7 @@
   number={6},
   pages={605--615},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/s10654-023-00997-3},
 }
 
@@ -65,7 +65,7 @@
   number={2},
   pages={179--187},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/s10654-022-00916-y},
 }
 
@@ -77,7 +77,7 @@
   volume={35},
   pages={601--611},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1007/s10654-020-00633-4},
 }
 
