### 2024-04-27 #6454 @elimillera
> 👋 @elimillera - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@crvernon I would be able to review in the next 2-3 weeks.
