### 2024-04-28 #6454 @ZekeMarshall
> 👋 @ZekeMarshall - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon , yes I'm happy to help! I can work through this next week and hopefully have my checklist completed by next Friday.
