### 2024-04-29 #6693 @editorialbot
[REVIEW]: DataRepExp: a R shiny Application that makes Data FAIR for Data Repositories
**Submitting author:** @RoryChenXY (<a href="http://orcid.org/0000-0002-2094-2682">Xinyue (Rory) Chen</a>)
**Repository:** https://github.com/RoryChenXY/DataRepExp_public
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
