### 2024-05-06 #6693 @ZekeMarshall
> > Hi @RoryChenXY, I'm just working through my review checklist now. Have you considered making versioned releases of DataRepExp_public? One other question, is there a "DataRepExp_private" repository? If so do you plan to keep future developments synchronised between these repositories?
>
> Hi @ZekeMarshall. 1. I think making versioned releases would be great; 2. There is a private repository that uses some AWS services and links to the sensitive data we use, with domain-specific variables. In the future, if we add more modules/functions in the private repository, I will update the public one manually. Due to security reasons, no synchronisation would happen between repositories.
>
> Thanks, Rory

Hi @RoryChenXY ,

That sounds good! Before I tick off the "Reproducibility" checklist item I would prefer that there was an initial versioned release.

Cheers,

Zeke
