### 2024-05-06 #6693 @ZekeMarshall
Hi @RoryChenXY , I don't think I can tick the "Automated tests" checklist item, as there are no tests in the repository. I think that all functions in the repository should be tested, which in DataRepExp consist of the functions in the vis_functions.R file. You could also test your modules, though I don't think that's strictly necessary. Best regards, Zeke
