### 2024-05-06 #6693 @ZekeMarshall
Hi @RoryChenXY , to tick of the "Installation" checklist item I think the repository requires dependencies to be recorded, i'd use [{renv}](https://rstudio.github.io/renv/articles/renv.html) to do this. Best regards, Zeke
