### 2024-05-07 #6693 @pydemull
Hi @RoryChenXY , regarding co-authorship, while JOSS does not require mentioning the kinds of the contributions from the different co-authors, I think it could be good to mention somewhere (eg, in the footer of the app or on a dedicated page of the app, or in the repo) what have been these kinds of contributions (eg, XX, YY and ZZ, ..., conceived the app (design, aim, features, etc), and XX and YY developed the app (the code)). A potential interest could be to provide some clues about the confidence a user can have in the app (especially if it is a new piece of software that has not been fully scrutinized by the community) : the more there are people who have the skills to check the code, the less there are risks of undetected bugs (I think).  I am not saying that a single person is not able to provide perfectly correct code of course. It is not mandatory so I check the box for this item.
