### 2024-05-23 #6693 @RoryChenXY
Dear All,

For better reproducibility, I have been experimenting with 'testthat' and unit testing, and realised my work needs to be improved so it can be used by others. This is the first time I am sharing my code with others.

I think the best way is to organise the app as a package.
I am reading https://r-pkgs.org/ and https://engineering-shiny.org/ at the moment.

Apologise for the delays, but there is a learning curve. I will do my best.

Please let me know if you don't think the package is necessary.

Thanks,
Rory
