### 2024-05-23 #6693 @crvernon
:wave: @RoryChenXY - first, I am so happy that you are sharing your code for the first time!  Our goal here at JOSS is to make sure we help you put together the best product possible.

As far as the tests go, I am glad you are working on these.  They will make your code more reliable and stable.

Your reviewers @elimillera, @pydemull, and @ZekeMarshall can provide some solid feedback and help as you go through this process.  Don't hesitate to reach out!  Thanks!
