+++ datarepexp/inst/golem-config.yml
@@ -1,0 +1,11 @@
+default:
+  golem_name: datarepexp
+  golem_version: 0.0.0.9000
+  app_prod: no
+
+production:
+  app_prod: yes
+
+dev:
+  golem_wd: !expr golem::pkg_path()
+
