+++ datarepexp/dev/01_start.R
@@ -20,11 +20,12 @@
 ##
 golem::fill_desc(
   pkg_name = "datarepexp", # The Name of the package containing the App
-  pkg_title = "PKG_TITLE", # The Title of the package containing the App
-  pkg_description = "PKG_DESC.", # The Description of the package containing the App
-  author_first_name = "AUTHOR_FIRST", # Your First Name
-  author_last_name = "AUTHOR_LAST", # Your Last Name
-  author_email = "AUTHOR@MAIL.COM", # Your Email
+  pkg_title = "Data Repository Explorer", # The Title of the package containing the App
+  pkg_description = "The Data Repository Explorer, is an open-source R Shiny application developed to improve the
+                     findability, accessibility, interoperability, and reusability (FAIR) of research data held in a data repository.", # The Description of the package containing the App
+  author_first_name = "Rory", # Your First Name
+  author_last_name = "Chen", # Your Last Name
+  author_email = "rory.chen.xy@gmail.com", # Your Email
   repo_url = NULL, # The URL of the GitHub Repo (optional),
   pkg_version = "0.0.0.9000" # The Version of the package containing the App
 )
