### 2024-05-27 3cf580d
Set {golem} options
## Set {golem} options ----
golem::set_golem_options()
