+++ datarepexp/NEWS.md
@@ -1,0 +1,3 @@
+# datarepexp (development version)
+
+* Initial CRAN submission.
