+++ datarepexp/README.Rmd
@@ -16,6 +16,7 @@ knitr::opts_chunk$set(
 # datarepexp
 
 <!-- badges: start -->
+[![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
 <!-- badges: end -->
 
 The goal of datarepexp is to ...
