### 2024-05-27 8aa2ff4
Create a template for tests:
golem::use_recommended_tests()
