+++ datarepexp/DESCRIPTION
@@ -11,7 +11,7 @@ Imports:
     shiny (>= 1.8.1.1)
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.1.1
+RoxygenNote: 7.3.1
 Suggests: 
     spelling,
     testthat (>= 3.0.0)
