+++ datarepexp/NAMESPACE
@@ -2,12 +2,14 @@
 
 export(run_app)
 import(shiny)
+import(shinydashboard)
 importFrom(golem,activate_js)
 importFrom(golem,add_resource_path)
 importFrom(golem,bundle_resources)
 importFrom(golem,favicon)
 importFrom(golem,with_golem_options)
 importFrom(shiny,HTML)
+importFrom(shiny,NS)
 importFrom(shiny,column)
 importFrom(shiny,shinyApp)
 importFrom(shiny,tagAppendAttributes)
