### 2024-05-27 d0ff86b
Update mod_tab1_statement.R
Tab 1 contents - just some statement
