### 2024-05-28 7861d53
Update mod_tab2_meta.R
module 2 - metadata table prototyping
