+++ datarepexp/data-raw/02_ppt_simulation.R
@@ -1,0 +1,269 @@
+#LIBRARY########################################################################
+library(dplyr) #data manipulation
+library(tidyverse) #data manipulation
+library(simstudy) #Generate simulated data
+library(data.table) #Data manipulation
+
+#SEED###########################################################################
+set.seed(951823) #Set seed for randomization
+
+#IMPORT FILES###################################################################
+load("data-raw/temp/simmeta.RData")
+
+ppt_dist <- setDT(read.csv("data-raw/source/ppt_Dist.csv", header = TRUE, fileEncoding="UTF-8-BOM"))#Self defined variable distribution
+ppt_Mdist <- setDT(read.csv("data-raw/source/ppt_MissDist.csv", header = TRUE, fileEncoding="UTF-8-BOM")) #defines the parameters of missing
+
+#FUNCTION#######################################################################
+#Assuming studies from same continent will follow similar distribution
+#I generate data for each continent first
+#Then use the function below to get data for each studies
+#Also added some validation - to make sure consistency between metadata and participant level data for each study
+Obs_Generator <- function(meta, Obsbank){
+  temp <- Obsbank[sample(.N, meta$STUDYSIZE)] #Grab n (Study sample size) from the databank
+  temp <- as.data.frame(temp)
+  minage <- meta$MINAGE
+  newage <- meta$MINAGE + round(abs(rnorm(meta$STUDYSIZE, mean = 0, sd=5))) #min age at recruitment + new age
+
+  temp <- temp %>%
+    mutate(YOD = na_if(YOD, DECEASED == 0)) %>% #if not dead, Year of death should be NA
+    mutate(YOD = case_when(DECEASED == 1 ~ YODtemp)) %>%
+    mutate(STUDY = meta$STUDY, #STUDY name
+           ID = paste(meta$STUDY, sprintf("%04d", 1:meta$STUDYSIZE), sep="")) %>%    #Set ID
+    mutate(AGEATASS = case_when(min(AGEATASS) <  minage ~ newage, #Make sure Age at assessment >= Study min age at recruitment
+                                min(AGEATASS) >= minage ~ AGEATASS)) %>%
+    mutate(SCALE4 = 0.3*AGEATASS + 0.2*ALCSTAT*AGEATASS + 5*abs(SCALE4e)) %>%
+    mutate(SCALE4 = case_when(SCALE4 <  100 ~ SCALE4, #Make sure Age at assessment >= Study min age at recruitment
+                              SCALE4 >= 100 ~ 100)) %>%
+    mutate(AGEATASS = round(AGEATASS), #Age round
+           SCALE1 = round(SCALE1), #Round -scales
+           SCALE2 = round(SCALE2),
+           SCALE3 = round(SCALE3),
+           SCALE4 = round(SCALE4)) %>%
+    relocate(c(STUDY,ID), .before = id) %>%
+    select(-c(id, YODtemp, SCALE1temp, SCALE4e))
+
+  if(meta$cat03 == 0){ #if no medical history data
+    temp$DIA1 <- NA
+    temp$DIA2 <- NA
+    temp$DIA3 <- NA
+    temp$DIA4 <- NA
+  }
[…]
