+++ datarepexp/dev/02_dev.R
@@ -37,7 +37,7 @@ golem::add_sass_file("custom")
 
 ## Add internal datasets ----
 ## If you have data in your package
-usethis::use_data_raw(name = "my_dataset", open = FALSE)
+usethis::use_data_raw(name = "03_data_prepare", open = FALSE)
 
 ## Tests ----
 ## Add one line by test you want to create

[…]
