### 2024-05-28 RoryChenXY/DataRepExp_public#2 @elimillera
JOSS Review: Dependency Managment 
From the JOSS Guidelines:

>  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

There does not appear to be a list of packages and their versions. The app.R file does contain the libraries that are loading in, and as of today I have no issues installing and loading the packages, however this may not be the case when updates occur to dependencies.

For a project like this I would recommend using {renv} to manage the dependencies of the application. This will create a machine-readable JSON file where environment state can be saved and restored for better reproducibility: https://rstudio.github.io/renv/articles/renv.html
