### 2024-05-28 RoryChenXY/DataRepExp_public#3 @elimillera
JOSS Review: Documentation 
From the JOSS Guidelines:

>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I would suggest adding how this could be used for non-simulated data and documentation of how this data should be structured. I'm not familiar with FAIR data so I'm not sure if the 'simdata.RData' is standard, but a note for how this Rdata file should be structured or a helper function that does the load from disk, would help a user know how to use their own data.
