### 2024-05-28 RoryChenXY/DataRepExp_public#4 @elimillera
JOSS Review: Contributing Guidelines
From the JOSS guidelines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I would suggest using this generator to create a CONTRIBUTING.md file in your repository to meet this requirement: https://generator.contributing.md/
