### 2024-05-28 #6693 @elimillera
@crvernon My check is complete. I have opened the following issues:
- [ ] https://github.com/RoryChenXY/DataRepExp_public/issues/2
- [ ] https://github.com/RoryChenXY/DataRepExp_public/issues/3
- [ ] https://github.com/RoryChenXY/DataRepExp_public/issues/4

As noted by the other editors, some tests would also be good to add for the application, generally I've seen this done in package framework, however that isn't a requirement and just a testthat directory should work without the package framework.

Nice work @RoryChenXY! Let me know if you would like me to expand on any of the points above.
