### 2024-05-28 b5da379
Update app_server.R
load tab 2 module
