### 2024-05-28 5f22214
Create mod_tab3_filter.R
create module file for tab 3
