### 2024-05-30 e3653ad
Create mod_tab3_filter_utils_inputs.R
A checkbox group input with only yes/no options
