+++ datarepexp/R/mod_tab3_filter_utils_inputs.R
@@ -18,7 +18,7 @@
 #' )
 #'
 #' server <- function(input, output, session) {
-#'   output$value <- renderText({ input$somevalue })
+#'   output$value <- renderText({input$somevalue})
 #' }
 #'
 #' shinyApp(ui, server)
@@ -33,5 +33,43 @@ yn_inputs <- function(inputId, label) {
     label = label,
     choices = c("No", "Yes"),
     selected = c("No", "Yes")
+  )
+}
+
+#' @title Check Box Group Input with only No/Yes/Missing Options
+#'
+#' @description
+#' Create a check box group input with only No/Yes/Missing Options
+#'
+#' @param inputId Input ID
+#' @param label Input label
+#'
+#' @return A list of HTML elements that can be added to a UI definition.
+#'
+#' @examples
+#' ## Only run examples in interactive R sessions
+#' if (interactive()) {
+#'
+#' ui <- fluidPage(
+#'   ynm_inputs("somevalue", "A Yes or No or Missing checkbox"),
+#'   verbatimTextOutput("value")
+#' )
+#'
+#' server <- function(input, output, session) {
+#'   output$value <- renderText({input$somevalue})
+#' }
+#'
+#' shinyApp(ui, server)
+#' }
+#'
+#' @importFrom shiny checkboxGroupInput
+#'
+#' @export
+ynm_inputs <- function(inputId, label) {
+  shiny::checkboxGroupInput(
+    inputId = inputId,
+    label = label,
+    choices = c("No", "Yes", "Missing"),
+    selected = c("No", "Yes", "Missing")
   )
 }
