### 2024-05-30 b710578
Create mod_tab3_filter_utils_apply.R
functions to apply filter
