+++ datarepexp/man/fct_inputs.Rd
@@ -1,0 +1,40 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mod_tab3_filter_utils_inputs.R
+\name{fct_inputs}
+\alias{fct_inputs}
+\title{Check Box Group Input with choices of the levels of a factor variable.}
+\usage{
+fct_inputs(inputId, label, df, dfvname)
+}
+\arguments{
+\item{inputId}{Input ID}
+
+\item{label}{Input label}
+
+\item{df}{A data frame with the factor variable}
+
+\item{dfvname}{A factor variable name}
+}
+\value{
+A list of HTML elements that can be added to a UI definition.
+}
+\description{
+Create a check box group input with choices of the levels of a factor variable.
+}
+\examples{
+## Only run examples in interactive R sessions
+if (interactive()) {
+dfsample <- data.frame(fctvar = as.factor(sample(letters[1:3], 10, replace = TRUE)))
+
+ui <- fluidPage(
+  fct_inputs("somevalue", "A factor checkbox", dfsample, fctvar),
+  verbatimTextOutput("value")
+)
+
+server <- function(input, output, session) {
+  output$value <- renderText({input$somevalue})
+}
+
+shinyApp(ui, server)
+}
+}
