+++ datarepexp/man/apply_slide_filter.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mod_tab3_filter_utils_apply.R
+\name{apply_slide_filter}
+\alias{apply_slide_filter}
+\title{Applying filters to numeric variable based on the slider inputs}
+\usage{
+apply_slide_filter(df, col, vals)
+}
+\arguments{
+\item{df}{A data frame}
+
+\item{col}{A column name, of the numeric variable}
+
+\item{vals}{A list of values, min, max, include missing}
+}
+\value{
+A filtered data frame
+}
+\description{
+Applying filters to numeric variable based on the slider inputs
+}
+\examples{
+dfsample <- data.frame(numvar = sample(1:10, 100, replace = TRUE))
+df_filtered <- apply_slide_filter(dfsample, 'numvar', list(3, 8, TRUE))
+
+}
