+++ datarepexp/R/mod_tab3_filter.R
@@ -4,28 +4,596 @@
 #'
 #' @param id,input,output,session Internal parameters for {shiny}.
 #'
-#' @noRd 
+#' @noRd
 #'
-#' @importFrom shiny NS tagList 
-mod_tab3_filter_ui <- function(id){
+#' @import dplyr
+#' @import DT
+#' @import shiny
+#' @importFrom magrittr %>%
+#' @importFrom purrr map map2 reduce2 set_names
+#' @importFrom shinyjs useShinyjs reset
+#' @importFrom useful compare.list
+
+# DT table option for Summary Tables Tab
+dtoptions2 <- list(
+  searching = FALSE,
+  scrollX = TRUE,
+  dom = 'Bfrtip',
+  buttons = list('pageLength',
+                 list(
+                   extend = 'colvis',
+                   columns = ":gt(0)"),
+                 'colvisRestore',
+                 'csv',
+                 'copy',
+                 'print')
+)
+
+mod_tab3_filter_ui <- function(id, metadf, pptdf) {
   ns <- NS(id)
   tagList(
- 
+    shinyjs::useShinyjs(), # Java Script for reset
+    div(
+      id = NS(id, "allfilters"),
+      htmlOutput(NS(id, "tab3statement")),
+      fluidRow( # Reset Button for all filters
+        column(1,
+          offset = 10, align = "right",
+          div(
+            style = "margin-bottom: 20px;",
+            actionButton(
+              inputId = NS(id, "resetallf"), # Reset All Filters
+              label = "Clear All Filters",
+              icon = icon("fas fa-sync")
+            )
+          )
+        )
+      ),
+      tabsetPanel(
+        id = NS(id, "filtertabs"),
+        type = "tabs",
+        tabPanel(
+          "Study Filters", ## Start of Study Filters Tab #####################################
+          div(
+            id = NS(id, "sfilters"),
+            fluidRow(br()),
+            fluidRow(
+              column(1,
+                offset = 10, align = "right",
+                div(
+                  style = "margin-bottom: 20px;",
+                  actionButton(
+                    inputId = NS(id, "resetsf"), # Reset Study Filters
+                    label = "Clear Study Filters",
+                    icon = icon("fas fa-sync")
+                  )
+                )
+              )
+            ),
+            wellPanel( #### Study Metadata #####################################
+              h3("Study metadata"),
+              fluidRow( # study name and continent
+                column(6, wellPanel(
+                  shinyWidgets::pickerInput(
+                    inputId = NS(id, "study"),
+                    label = "Studies Selected",
+                    choices = list(
+                      "Africa"        = as.character(metadf$STUDY[metadf$CONTINENT == "Africa"]),
+                      "Asia"          = as.character(metadf$STUDY[metadf$CONTINENT == "Asia"]),
+                      "Europe"        = as.character(metadf$STUDY[metadf$CONTINENT == "Europe"]),
+                      "North America" = as.character(metadf$STUDY[metadf$CONTINENT == "North America"]),
+                      "Oceania"       = as.character(metadf$STUDY[metadf$CONTINENT == "Oceania"]),
+                      "South America" = as.character(metadf$STUDY[metadf$CONTINENT == "South America"])
+                    ),
+                    selected = as.character(metadf$STUDY),
+                    multiple = TRUE,
+                    options = list(
+                      `actions-box` = TRUE,
+                      `selected-text-format` = "count > 3",
+                      `count-selected-text` = "{0} Studies selected",
+                      `none-selected-text` = "Select Studies",
+                      `live-search` = TRUE,
+                      `select-all` = TRUE,
+                      `deselect-all` = TRUE
+                    )
+                  )
+                )),
+                column(6, wellPanel(
[…]
