+++ datarepexp/NAMESPACE
@@ -1,17 +1,37 @@
 # Generated by roxygen2: do not edit by hand
 
+export(apply_factor_filter)
+export(apply_slide_filter)
+export(fct_inputs)
 export(run_app)
+export(yn_inputs)
+export(ynm_inputs)
+import(DT)
+import(dplyr)
 import(shiny)
 import(shinydashboard)
+importFrom(dplyr,filter)
 importFrom(golem,activate_js)
 importFrom(golem,add_resource_path)
 importFrom(golem,bundle_resources)
 importFrom(golem,favicon)
 importFrom(golem,with_golem_options)
+importFrom(magrittr,"%>%")
+importFrom(purrr,map)
+importFrom(purrr,map2)
+importFrom(purrr,reduce2)
+importFrom(purrr,set_names)
+importFrom(rlang,"!!")
+importFrom(rlang,sym)
 importFrom(shiny,HTML)
 importFrom(shiny,NS)
+importFrom(shiny,checkboxGroupInput)
 importFrom(shiny,column)
 importFrom(shiny,shinyApp)
 importFrom(shiny,tagAppendAttributes)
 importFrom(shiny,tagList)
 importFrom(shiny,tags)
+importFrom(shiny,wellPanel)
+importFrom(shinyjs,reset)
+importFrom(shinyjs,useShinyjs)
+importFrom(useful,compare.list)
