### 2024-05-30 cb633ed
Update app_server.R
module 3 included in server
