### 2024-05-31 a24533b
Update mod_tab3_filter.R
Added a table for filter report
