### 2024-06-03 d53e8cb
Update mod_tab2_meta.R
Warning: replacing previous import 'DT::dataTableOutput' by 'shiny::dataTableOutput'

update the namespace info to address the warning
