### 2024-06-03 51bfe7f
Update mod_tab3_filter.R
Warning: replacing previous import 'DT::dataTableOutput' by 'shiny::dataTableOutput'

update the namespace info to address the warning
