### 2024-06-03 3e7f6e8
Update mod_tab3_filter_utils_apply.R
minor updates for example
