+++ datarepexp/man/mod_tab3_filter_ui.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/mod_tab3_filter.R
+\name{mod_tab3_filter_ui}
+\alias{mod_tab3_filter_ui}
+\title{tab3_filter UI Function}
+\usage{
+mod_tab3_filter_ui(id, metadf, pptdf)
+}
+\description{
+tab3_filter UI Function
+}
