### 2024-06-04 7cc6b9d
Create utils_countfactor.R
count factor levels
