### 2024-06-04 ba937af
Create test-utils_countfactor.R
test countfactor
