### 2024-06-06 8e1e0e4
Update utils_countfactor.R
minor
