+++ datarepexp/R/fct_visplotly.R
@@ -1,0 +1,145 @@
+#' @title plotly_pie
+#'
+#' @description A standardised function to generate a plotly pie chart
+#'
+#' @param data A data frame to be processed
+#' @param var A variable to generate pie chart
+#'
+#' @importFrom magrittr %>%
+#' @importFrom rlang enquo
+#' @importFrom dplyr mutate count
+#' @importFrom grDevices hcl.colors
+#' @importFrom plotly plot_ly layout
+#'
+#' @return A plotly pie chart.
+#'
+#' @export
+#'
+#' @examples
+#' data <- data.frame(fctvar = as.factor(rep(c("a", "b", "c", "c"), 10)))
+#' plotly_pie(data, fctvar)
+plotly_pie <- function(data, var){
+
+  # Check if Data is a data.frame
+  if (class(data) != "data.frame") {
+    stop("Expecting a data frame")
+  }
+
+  # Capture the variable name as a quosure for use in dplyr functions
+  var <- rlang::enquo(var)
+  # Create a new data frame with counts for each level and a new column for the variable name
+  df <- data %>%
+    dplyr::count(!!(var), name = "freq") %>% # count unique values
+    dplyr::mutate(var_level = !!(var)) # column variable level
+
+  # Create a colour palette based on the number of levels of the variable
+  cpalette <- grDevices::hcl.colors(nrow(df), palette = "Pastel1")
+
+  # Create a pie chart using plotly
+  p <- plotly::plot_ly(df, labels = ~var_level, values = ~freq, type = "pie",
+                       sort = FALSE,
+                       textinfo = "label+percent", # Display the label and percentage for each slice
+                       textposition = "inside",
+                       insidetextfont = list(color = "#FFFFFF"),
+                       hoverinfo = "text", # Display the text string on hover
+                       text = ~paste("Count:", freq),# Set the hover text
+                       marker = list(colors = cpalette, # Assign colors to each slice
+                                     line = list(color = "#FFFFFF", width = 1)),
+                       showlegend = TRUE) %>%
+    plotly::layout(legend = list(orientation = "h")) # Set the legend to horizontal
+
+  return(p)
+}
+
+
+#' @title plotly_bar_multifct
+#'
+#' @description A function to generate a plotly bar chart that combined multiple factors with same levels.
+#'
+#' @param data A data frame generated with [countfactor] function, including 4 variables var_label, var_level, freq, and percent
+#'
+#' @importFrom ggplot2 ggplot aes geom_bar geom_text coord_flip scale_fill_brewer theme
+#' @importFrom plotly ggplotly layout
+#'
+#' @return A plotly bar chart for multiple factors with same level.
+#'
+#' @export
+#'
+#' @examples
+#' bar_multifct_df <- data.frame(var_label = c("AA", "AA", "BB", "BB", "CC","CC", "DD", "DD"),
+#'                               var_level = as.factor(rep(c("No", "Yes"), 4)),
+#'                               freq      = c(0, 10, 3, 7, 2, 8, 4, 6),
+#'                               percent   = c("0%", "100%", "30%", "70%","20%", "80%", "40%", "60%"))
+#' plotly_bar_multifct(bar_multifct_df)
+plotly_bar_multifct <- function(data){
+  # Used with the countfactor Function to produce combined bar charts for multiple factors
+
+  if (all(names(data) != c("var_label", "var_level", "freq", "percent"))) {
+    stop("Expecting a data frame with following variables: var_label, var_level, freq, and percent")
+  }
+
+  p <- ggplot2::ggplot(data,
+                       ggplot2::aes(x = var_label,
+                                    y = freq,
[…]
