### 2024-06-06 e67e36f
Create countfactor.Rd
RD by roxygen2
