### 2024-06-06 f74f5ca
Create mod_tab4_visual.R
module for tab 4
