+++ datarepexp/R/mod_tab4_visual.R
@@ -1,0 +1,522 @@
+#' tab4_visual UI Function
+#'
+#' @description A shiny Module for tab 4 visualisation.
+#'
+#' @param id,input,output,session Internal parameters for {shiny}.
+#'
+#' @noRd
+#'
+#' @import shiny
+#' @import ggplot2
+#' @importFrom plotly plotlyOutput renderPlotly ggplotly
+#' @importFrom magrittr %>%
+#' @importFrom dplyr mutate select filter summarise group_by ungroup n
+#' @importFrom tidyr complete drop_na
+#' @importFrom scales breaks_pretty
+#' @importFrom forcats fct_rev
+
+
+# tab4_visual UI Function
+mod_tab4_visual_ui <- function(id){
+  ns <- NS(id)
+  tagList(
+    h2("The Visualisation Tab generates results
+        based on the filters you have applied."),
+    fluidPage(
+      uiOutput(NS(id, "fvalid")),
+      tabsetPanel(
+        id = NS(id, "visaultabs"),
+        type = "tabs",
+        tabPanel("Metadata", uiOutput(NS(id, "tab4meta"))),
+        tabPanel("Demographics", uiOutput(NS(id, "tab4dem"))),
+        tabPanel("Lifestyle", uiOutput(NS(id, "tab4life"))),
+        tabPanel("Scales", uiOutput(NS(id, "tab4scale"))),
+        tabPanel("Health and Family History", uiOutput(NS(id, "tab4health"))),
+        tabPanel("Imaging and Genomic Data", uiOutput(NS(id, "tab4imgeno")))
+      )
+    )
+
+  )
+}
+
+#' tab4_visual Server Functions
+#'
+#' @noRd
+mod_tab4_visual_server <- function(id, metadf, react_df){
+  moduleServer( id, function(input, output, session){
+    ns <- session$ns
+    # Check the filtered results
+    output$fvalid <- renderUI({
+      if (nrow(react_df()) == 0) {
+        h2("Base on the filters you have applied,
+            no participants were identified, please update your selection.")
+      } else {
+        a <- react_df()
+        df <- a %>% count(SEX, .drop = FALSE, name = "Count")
+        fluidPage(
+          h3("Base on the filters you applied,
+              you have found: "),
+          fluidRow(
+            column(width = 2, HTML(paste(icon("object-ungroup"), "Studies: ", length(unique(a$STUDY))))),
+            column(width = 2, HTML(paste(icon("users"), "Participants: ", nrow(a)))),
+            column(width = 2, HTML(paste(icon("male"), "Male: ", df$Count[1]))),
+            column(width = 2, HTML(paste(icon("female"), "Female: ", df$Count[2]))),
+            column(width = 2, HTML(paste(icon("transgender"), "Other: ", df$Count[3]))),
+            column(width = 2, HTML(paste(icon("question-circle"), "Sex Missing: ", df$Count[4])))
+          ),
+        )
+      }
+    })
+
+    ## Tab 4-1 Metadata ########################################################
+    ###  Tab 4-1 Layout ########################################################
+    output$tab4meta <- renderUI({
+      if (nrow(react_df()) == 0) {
+        h2("Please update the your filter selection.")
+      } else {
+        fluidPage(
+          fluidRow(
+            column(6, h4("Min Age at Recruitment"), plotly::plotlyOutput(NS(id, "minagebar"))),
+            column(6, h4("Sample Size"), plotly::plotlyOutput(NS(id, "sizebar")))
+          ),
+          fluidRow(
+            column(3, h4("Data Available through Repository"), plotly::plotlyOutput(NS(id, "avapie"))),
[…]
