+++ datarepexp/NAMESPACE
@@ -2,27 +2,62 @@
 
 export(apply_factor_filter)
 export(apply_slide_filter)
+export(countfactor)
 export(fct_inputs)
+export(plotly_bar_multifct)
+export(plotly_histogram_grouped)
+export(plotly_pie)
 export(run_app)
 export(yn_inputs)
 export(ynm_inputs)
-import(DT)
-import(dplyr)
+import(ggplot2)
 import(shiny)
+import(shiny, except=c(dataTableOutput, renderDataTable))
 import(shinydashboard)
+importFrom(DT,DTOutput)
+importFrom(DT,clearSearch)
+importFrom(DT,dataTableProxy)
+importFrom(DT,datatable)
+importFrom(DT,renderDT)
+importFrom(dplyr,arrange)
+importFrom(dplyr,count)
 importFrom(dplyr,filter)
+importFrom(dplyr,group_by)
+importFrom(dplyr,mutate)
+importFrom(dplyr,n)
+importFrom(dplyr,select)
+importFrom(dplyr,summarise)
+importFrom(dplyr,ungroup)
+importFrom(forcats,fct_rev)
+importFrom(ggplot2,aes)
+importFrom(ggplot2,coord_flip)
+importFrom(ggplot2,geom_bar)
+importFrom(ggplot2,geom_text)
+importFrom(ggplot2,ggplot)
+importFrom(ggplot2,scale_fill_brewer)
+importFrom(ggplot2,theme)
 importFrom(golem,activate_js)
 importFrom(golem,add_resource_path)
 importFrom(golem,bundle_resources)
 importFrom(golem,favicon)
 importFrom(golem,with_golem_options)
+importFrom(grDevices,hcl.colors)
 importFrom(magrittr,"%>%")
+importFrom(plotly,add_histogram)
+importFrom(plotly,ggplotly)
+importFrom(plotly,layout)
+importFrom(plotly,plot_ly)
+importFrom(plotly,plotlyOutput)
+importFrom(plotly,renderPlotly)
 importFrom(purrr,map)
 importFrom(purrr,map2)
 importFrom(purrr,reduce2)
 importFrom(purrr,set_names)
 importFrom(rlang,"!!")
+importFrom(rlang,enquo)
 importFrom(rlang,sym)
+importFrom(scales,breaks_pretty)
+importFrom(scales,percent)
 importFrom(shiny,HTML)
 importFrom(shiny,NS)
 importFrom(shiny,checkboxGroupInput)
@@ -32,6 +67,9 @@ importFrom(shiny,tagAppendAttributes)
 importFrom(shiny,tagList)
 importFrom(shiny,tags)
 importFrom(shiny,wellPanel)
+importFrom(shinyWidgets,pickerInput)
 importFrom(shinyjs,reset)
 importFrom(shinyjs,useShinyjs)
+importFrom(tidyr,complete)
+importFrom(tidyr,drop_na)
 importFrom(useful,compare.list)
