### 2024-06-07 b9102ee
Update mod_tab4_visual.R
update plotly_histogram_grouped, adding plotly buttons to change barmode; adding labels
