### 2024-06-07 3227009
Create mod_tab5_pac.R
module  Preliminary Analysis - Categorical Outcome
