+++ datarepexp/R/mod_tab5_pac.R
@@ -1,0 +1,293 @@
+#' tab5_pac UI Function
+#'
+#' @description A shiny Module.
+#'
+#' @param id,input,output,session Internal parameters for {shiny}.
+#'
+#' @noRd
+#'
+#' @import shiny
+#' @importFrom shinyWidgets pickerInput
+#' @importFrom magrittr %>%
+#' @importFrom dplyr count mutate
+#' @importFrom tidyr drop_na pivot_wider
+#' @importFrom RColorBrewer brewer.pal
+#' @importFrom plotly plotlyOutput renderPlotly ggplotly plot_ly layout
+
+mod_tab5_pac_ui <- function(id){
+  ns <- NS(id)
+  quanchoice <- list(
+    "Age at Assessment",
+    "Year of Death",
+    "BMI",
+    "Scale 1",
+    "Scale 2",
+    "Scale 3",
+    "Scale 4"
+  )
+
+  catevchoice <- list(
+    "Demographics"               = list("Ethnic Background", "Sex", "High School Educated", "Married/De-facto", "Deceased"),
+    "Lifestyle"                  = list("Smoking Status", "Alcohol Use Status"),
+    "Disease Diagnoses"          = list("Disease Diagnosis 1", "Disease Diagnosis 2", "Disease Diagnosis 3", "Disease Diagnosis 4"),
+    "Service Utilization"        = list("Hospital Outpatient", "Hospital Inpatient", "GP"),
+    "Family History of Diseases" = list("Family History of Diagnosis 1", "Family History of Diagnosis 2", "Family History of Diagnosis 3"),
+    "Imaging Data"               = list("MRI Collected", "Imaging 1 Collected", "Imaging 2 Collected"),
+    "Genomic Data"               = list("Geno type 1", "Geno type 2")
+  )
+  tagList(
+    h2("The Preliminary Analysis Tab generates results based on the filters you have applied."),
+    h2("Please allow up to a minute for the results to load."),
+    fluidPage(
+      fluidRow(
+        column(3, shinyWidgets::pickerInput(
+          inputId = NS(id, "ycate"),
+          label = "Categorical Y",
+          choices = catevchoice,
+          selected = "Disease Diagnosis 1",
+          multiple = FALSE
+        )),
+        column(3, selectInput(
+          inputId = NS(id, "cx1q"),
+          label = "Quantitative X1",
+          choices = quanchoice,
+          selected = "Scale 2",
+          multiple = FALSE
+        )),
+        column(3, shinyWidgets::pickerInput(
+          inputId = NS(id, "cx2c"),
+          label = "Categorical X2",
+          choices = catevchoice,
+          selected = "Alcohol Use Status",
+          multiple = FALSE
+        ))
+      ),
+      uiOutput(NS(id, "pac_valid")),
+      tabsetPanel(
+        id = NS(id, "pactabs"),
+        type = "tabs",
+        tabPanel("Univariate", uiOutput(NS(id, "pacuni"))),
+        tabPanel("Y~X1", uiOutput(NS(id, "pacx1"))),
+        tabPanel("Y~X2", uiOutput(NS(id, "pacx2")))
+      )
+    )
+  )
+}
+
+#' tab5_pac Server Functions
+#'
+#' @noRd
+mod_tab5_pac_server <- function(id, react_df){
+  moduleServer(id, function(input, output, session){
+    ns <- session$ns
+
+    # Check the filtered results and selected variables
+    output$pac_valid <- renderUI({
+      if (nrow(react_df()) == 0) {
+        h2("Base on the filters you applied,
+            no participants were identified, please update your selection.")
+      } else {
+        if (input$ycate == input$cx2c) {
+          h2("Please choose a different Categorical X2") # Variable selection validation
+        } else {
+          h3(paste0(
[…]
