### 2024-06-07 2b9c6f1
Create utils_selectedvar.R
return variable name based on variable label
