### 2024-06-07 c017d2f
Update fct_visplotly.R
add buttons for plotly histogram
