+++ datarepexp/man/selectedvar.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils_selectedvar.R
+\name{selectedvar}
+\alias{selectedvar}
+\title{selectedvar}
+\usage{
+selectedvar(var_label)
+}
+\arguments{
+\item{var_label}{variable label}
+}
+\value{
+variable name as a string
+}
+\description{
+Return variable name based on the variable label
+}
+\examples{
+vlabel <- "Year of Death"
+vname <- selectedvar(vlabel)
+}
