+++ datarepexp/man/plotly_histogram_grouped.Rd
@@ -4,14 +4,14 @@
 \alias{plotly_histogram_grouped}
 \title{plotly_histogram_grouped}
 \usage{
-plotly_histogram_grouped(df, var, groupvar, label)
+plotly_histogram_grouped(df, var, groupvar, varlabel, grouplabel)
 }
 \arguments{
 \item{df}{A data frame to be processed}
 
 \item{var}{A numeric variable to generate histogram}
 
-\item{groupvar}{Grouping variable}
+\item{groupvar}{A grouping variable}
 
 \item{label}{Legend label}
 }
@@ -24,4 +24,5 @@ A standardised function to generate a grouped histogram
 \examples{
 group_his_df <- data.frame(numvar = rep(1: 20, 5), fctvar = as.factor(c(rep(c("A", "B", "C"), 30), rep("A", 10))))
 plotly_histogram_grouped(df = group_his_df, var = "numvar", groupvar = "fctvar", label = "Groups")
+
 }
