+++ datarepexp/NAMESPACE
@@ -8,6 +8,7 @@ export(plotly_bar_multifct)
 export(plotly_histogram_grouped)
 export(plotly_pie)
 export(run_app)
+export(selectedvar)
 export(yn_inputs)
 export(ynm_inputs)
 import(ggplot2)
@@ -19,6 +20,7 @@ importFrom(DT,clearSearch)
 importFrom(DT,dataTableProxy)
 importFrom(DT,datatable)
 importFrom(DT,renderDT)
+importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,arrange)
 importFrom(dplyr,count)
 importFrom(dplyr,filter)
@@ -72,4 +74,5 @@ importFrom(shinyjs,reset)
 importFrom(shinyjs,useShinyjs)
 importFrom(tidyr,complete)
 importFrom(tidyr,drop_na)
+importFrom(tidyr,pivot_wider)
 importFrom(useful,compare.list)
