### 2024-06-07 4ca6baa
Update app_server.R
add mod_tab5_pac related info
