### 2024-06-12 81a4b56
Update run_app.R
minor changes
