### 2024-06-12 b47b9e5
Update DESCRIPTION
update dependencies
