+++ datarepexp/DESCRIPTION
@@ -8,14 +8,33 @@ License: MIT + file LICENSE
 Imports: 
     config (>= 0.3.2),
     golem (>= 0.4.1),
-    shiny (>= 1.8.1.1)
+    dplyr (>= 1.1.4),
+    magrittr (>= 2.0.3),
+    purrr (>= 1.0.2),
+    rlang (>= 1.1.4),
+    tidyr (>= 1.3.1),
+    useful (>= 1.2.6.1),
+    shinydashboard (>= 0.7.2),
+    shinyWidgets (>= 0.8.6),
+    shinyjs (>= 2.1.0),
+    DT (>= 0.33),
+    ggplot2 (>= 3.5.1),
+    plotly (>= 4.10.4),
+    scales (>= 1.3.0),
+    collapse (>= 2.0.14),
+    forcats (>= 1.0.0),
+    grDevices (>= 4.4.0),
+    RColorBrewer (>= 1.1.3),
+    testthat (>= 3.2.1.1)
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.1
 Suggests: 
-    spelling,
-    testthat (>= 3.0.0)
+    knitr,
+    rmarkdown
 Config/testthat/edition: 3
 Language: en-US
 Depends: 
-    R (>= 2.10)
+    R (>= 4.4),
+    shiny (>= 1.8.1.1)
+VignetteBuilder: knitr
