+++ datarepexp/R/app_ui.R
@@ -7,15 +7,15 @@
 #' @noRd
 app_ui <- function(request) {
   tagList(
-    # Leave this function for adding external resources
-    golem_add_external_resources(),
     # Your application UI logic
     shinydashboard::dashboardPage(
       skin = "black",
-      shinydashboard::dashboardHeader(title = "Data Repository Explorer"),
+      shinydashboard::dashboardHeader(title = "Repository Explorer"),
       shinydashboard::dashboardSidebar(uiOutput("sidebarpanel")),
       shinydashboard::dashboardBody(uiOutput("body"))
-    )
+    ),
+    # Leave this function for adding external resources
+    golem_add_external_resources()
   )
 }
 
@@ -27,6 +27,7 @@ app_ui <- function(request) {
 #' @import shiny
 #' @import shinydashboard
 #' @importFrom golem add_resource_path activate_js favicon bundle_resources
+#' @export
 #' @noRd
 golem_add_external_resources <- function() {
   add_resource_path(
@@ -35,11 +36,8 @@ golem_add_external_resources <- function() {
   )
 
   tags$head(
-    favicon(),
-    bundle_resources(
-      path = app_sys("app/www"),
-      app_title = "datarepexp"
-    )
+    tags$link(rel="stylesheet", type="text/css", href="www/custom.css"),
+    golem::favicon(),
     # Add here other external resources
     # for example, you can add shinyalert::useShinyalert()
   )
