### 2024-06-12 28e5ac3
module app functions
minor - app function for each module
