### 2024-06-12 b4fc48d
Create mod_tab5_paq.R
module tab 5 -PAQ
