+++ datarepexp/R/mod_tab5_paq.R
@@ -1,0 +1,617 @@
+#' tab5_paq UI Function
+#'
+#' @description A shiny Module.
+#'
+#' @param id,input,output,session Internal parameters for {shiny}.
+#'
+#' @noRd
+#'
+#' @import shiny
+#' @import ggplot2
+#' @importFrom shinyWidgets pickerInput
+#' @importFrom magrittr %>%
+#' @importFrom dplyr count mutate group_by summarise
+#' @importFrom collapse fmin fmax
+#' @importFrom tidyr drop_na pivot_wider
+#' @importFrom RColorBrewer brewer.pal
+#' @importFrom plotly plotlyOutput renderPlotly ggplotly plot_ly layout add_boxplot
+mod_tab5_paq_ui <- function(id){
+  ns <- NS(id)
+  quanchoice <- list(
+    "Age at Assessment",
+    "Year of Death",
+    "BMI",
+    "Scale 1",
+    "Scale 2",
+    "Scale 3",
+    "Scale 4"
+  )
+
+  catevchoice <- list(
+    "Demographics"               = list("Ethnic Background", "Sex", "High School Educated", "Married/De-facto", "Deceased"),
+    "Lifestyle"                  = list("Smoking Status", "Alcohol Use Status"),
+    "Disease Diagnoses"          = list("Disease Diagnosis 1", "Disease Diagnosis 2", "Disease Diagnosis 3", "Disease Diagnosis 4"),
+    "Service Utilization"        = list("Hospital Outpatient", "Hospital Inpatient", "GP"),
+    "Family History of Diseases" = list("Family History of Diagnosis 1", "Family History of Diagnosis 2", "Family History of Diagnosis 3"),
+    "Imaging Data"               = list("MRI Collected", "Imaging 1 Collected", "Imaging 2 Collected"),
+    "Genomic Data"               = list("Geno type 1", "Geno type 2")
+  )
+  tagList(
+    h2("The Preliminary Analysis Tab generates results based on the filters you have applied."),
+    h2("Please allow up to a minute for the results to load."),
+    fluidPage(
+      fluidRow(
+        column(3, selectInput(
+          inputId = NS(id, "yquan"),
+          label = "Quantitative Y",
+          choices = quanchoice,
+          selected = "Scale 4",
+          multiple = FALSE
+        )),
+        column(3, selectInput(
+          inputId = NS(id, "qx1q"),
+          label = "Quantitative X1",
+          choices = quanchoice,
+          selected = "Age at Assessment",
+          multiple = FALSE
+        )),
+        column(3, shinyWidgets::pickerInput(
+          inputId = NS(id, "qx2c"),
+          label = "Categorical X2",
+          choices = catevchoice,
+          selected = "Alcohol Use Status",
+          multiple = FALSE
+        )),
+        column(3, shinyWidgets::pickerInput(
+          inputId = NS(id, "qx3c"),
+          label = "Categorical X3",
+          choices = catevchoice,
+          selected = "Deceased",
+          multiple = FALSE
+        ))
+      ),
+      uiOutput(NS(id, "paq_valid")),
+      tabsetPanel(
+        id = NS(id, "paqtabs"),
+        type = "tabs",
+        tabPanel("Univariate", uiOutput(NS(id, "paquni"))),
+        tabPanel("Y~X1+X2", uiOutput(NS(id, "paqx1x2"))),
+        tabPanel("Y~X2", uiOutput(NS(id, "paqx2"))),
+        tabPanel("Y~X2+X3", uiOutput(NS(id, "paqx2x3")))
+      )
+    )
+
+  )
+}
+
+#' tab5_paq Server Functions
+#'
+#' @noRd
+mod_tab5_paq_server <- function(id, react_df){
+  moduleServer(id, function(input, output, session){
[…]
