### 2024-06-12 d3e3e39
Update mod_tab3_filter_utils_apply.R
minor roxygen updates
