+++ datarepexp/NAMESPACE
@@ -4,6 +4,13 @@ export(apply_factor_filter)
 export(apply_slide_filter)
 export(countfactor)
 export(fct_inputs)
+export(golem_add_external_resources)
+export(mod_tab1_statement_app)
+export(mod_tab2_meta_app)
+export(mod_tab3_filter_app)
+export(mod_tab4_visual_app)
+export(mod_tab5_pac_app)
+export(mod_tab5_paq_app)
 export(plotly_bar_multifct)
 export(plotly_histogram_grouped)
 export(plotly_pie)
@@ -21,6 +28,8 @@ importFrom(DT,dataTableProxy)
 importFrom(DT,datatable)
 importFrom(DT,renderDT)
 importFrom(RColorBrewer,brewer.pal)
+importFrom(collapse,fmax)
+importFrom(collapse,fmin)
 importFrom(dplyr,arrange)
 importFrom(dplyr,count)
 importFrom(dplyr,filter)
@@ -45,6 +54,7 @@ importFrom(golem,favicon)
 importFrom(golem,with_golem_options)
 importFrom(grDevices,hcl.colors)
 importFrom(magrittr,"%>%")
+importFrom(plotly,add_boxplot)
 importFrom(plotly,add_histogram)
 importFrom(plotly,ggplotly)
 importFrom(plotly,layout)
@@ -60,14 +70,10 @@ importFrom(rlang,enquo)
 importFrom(rlang,sym)
 importFrom(scales,breaks_pretty)
 importFrom(scales,percent)
-importFrom(shiny,HTML)
 importFrom(shiny,NS)
 importFrom(shiny,checkboxGroupInput)
-importFrom(shiny,column)
 importFrom(shiny,shinyApp)
-importFrom(shiny,tagAppendAttributes)
 importFrom(shiny,tagList)
-importFrom(shiny,tags)
 importFrom(shiny,wellPanel)
 importFrom(shinyWidgets,pickerInput)
 importFrom(shinyjs,reset)
