### 2024-06-12 d5e9261
Update utils_countfactor.R
minor roxygen updates
