### 2024-06-12 7efa769
Update utils_selectedvar.R
minor roxygen updates
