+++ datarepexp/app.R
@@ -1,0 +1,9 @@
+# Launch the ShinyApp (Do not remove this comment)
+# To deploy, run: rsconnect::deployApp()
+# Or use the blue button on top of this file
+
+pkgload::load_all()
+
+options("golem.app.prod" = TRUE)
+
+datarepexp::run_app() # add parameters here (if any)
