### 2024-06-12 f67cf90
Create test-utils_selectedvar.R
test function
