+++ datarepexp/DESCRIPTION
@@ -31,7 +31,9 @@ LazyData: true
 RoxygenNote: 7.3.1
 Suggests: 
     knitr,
-    rmarkdown
+    rmarkdown,
+    devtools,
+    shinytest2
 Config/testthat/edition: 3
 Language: en-US
 Depends: 
