### 2024-06-12 0cb6dc4
Update utils_selectedvar.R
error
