+++ datarepexp/dev/02_dev.R
@@ -39,18 +39,16 @@ golem::add_fct("visplotly", with_test = TRUE)
 golem::add_utils("countfactor", with_test = TRUE)
 golem::add_utils("selectedvar", with_test = TRUE)
 
-
-
 ## Creates fct_* and utils_*
-golem::add_fct("helpers", with_test = TRUE)
-golem::add_utils("helpers", with_test = TRUE)
+#golem::add_fct("helpers", with_test = TRUE)
+#golem::add_utils("helpers", with_test = TRUE)
 
 ## External resources
 ## Creates .js and .css files at inst/app/www
-golem::add_js_file("script")
-golem::add_js_handler("handlers")
+#golem::add_js_file("script")
+#golem::add_js_handler("handlers")
 golem::add_css_file("custom")
-golem::add_sass_file("custom")
+#golem::add_sass_file("custom")
 
 ## Add internal datasets ----
 ## If you have data in your package
