+++ datarepexp/vignettes/datarepexp.Rmd
@@ -1,0 +1,19 @@
+---
+title: "datarepexp"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{datarepexp}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(datarepexp)
+```
