+++ datarepexp/dev/04_renv.R
@@ -1,0 +1,3 @@
+library(renv)
+renv::init()
+renv::snapshot()
