+++ datarepexp/NAMESPACE
@@ -7,6 +7,7 @@ export(fct_inputs)
 export(golem_add_external_resources)
 export(mod_tab1_statement_app)
 export(mod_tab2_meta_app)
+export(mod_tab2_meta_server)
 export(mod_tab3_filter_app)
 export(mod_tab4_visual_app)
 export(mod_tab5_pac_app)
@@ -14,7 +15,7 @@ export(mod_tab5_paq_app)
 export(plotly_bar_multifct)
 export(plotly_histogram_grouped)
 export(plotly_pie)
-export(run_app)
+export(repexp_app)
 export(selectedvar)
 export(yn_inputs)
 export(ynm_inputs)
