+++ datarepexp/R/fct_visplotly.R
@@ -119,7 +119,7 @@ plotly_bar_multifct <- function(data){
 #'
 #' @examples
 #' group_his_df <- data.frame(numvar = rep(1: 20, 5), fctvar = as.factor(c(rep(c("A", "B", "C"), 30), rep("A", 10))))
-#' plotly_histogram_grouped(df = group_his_df, var = "numvar", groupvar = "fctvar", label = "Groups")
+#' plotly_histogram_grouped(df = group_his_df, var = "numvar", groupvar = "fctvar", varlabel = "Numbers", grouplabel = "Groups")
 #'
 plotly_histogram_grouped <- function(df, var, groupvar, varlabel, grouplabel){
 
