### 2024-06-14 c791743
Update plotly_histogram_grouped.Rd
example update
