+++ datarepexp/man/plotly_histogram_grouped.Rd
@@ -23,6 +23,6 @@ A standardised function to generate a grouped histogram with three display optio
 }
 \examples{
 group_his_df <- data.frame(numvar = rep(1: 20, 5), fctvar = as.factor(c(rep(c("A", "B", "C"), 30), rep("A", 10))))
-plotly_histogram_grouped(df = group_his_df, var = "numvar", groupvar = "fctvar", label = "Groups")
+plotly_histogram_grouped(df = group_his_df, var = "numvar", groupvar = "fctvar", varlabel = "Numbers", grouplabel = "Groups")
 
 }
