### 2024-06-14 3ea4e56
Update app.R
run the app function
