+++ datarepexp/app.R
@@ -2,8 +2,10 @@
 # To deploy, run: rsconnect::deployApp()
 # Or use the blue button on top of this file
 
-pkgload::load_all()
+devtools::load_all()
 
 options("golem.app.prod" = TRUE)
 
-datarepexp::run_app() # add parameters here (if any)
+library(datarepexp)
+
+datarepexp::repexp_app() # add parameters here (if any)
