+++ datarepexp/DESCRIPTION
@@ -24,16 +24,18 @@ Imports:
     collapse (>= 2.0.14),
     forcats (>= 1.0.0),
     grDevices (>= 4.4.0),
-    RColorBrewer (>= 1.1.3),
-    testthat (>= 3.2.1.1)
+    RColorBrewer (>= 1.1.3)
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.1
 Suggests: 
+    testthat (>= 3.2.1.1),
     knitr,
     rmarkdown,
     devtools,
-    shinytest2
+    shinytest2,
+    covr,
+    covrpage
 Config/testthat/edition: 3
 Language: en-US
 Depends: 
