### 2024-06-14 fb020c4
Update test-shinytest2.R
Test - applied filter: Asia study + Male + Current smoker
