+++ datarepexp/tests/testthat/test-shinytest2.R
@@ -794,3 +794,301 @@ test_that("{shinytest2} recording: datarepexp", {
   app$set_inputs(`plotly_hover-A` = character(0), allow_no_input_binding_ = TRUE,
       priority_ = "event")
 })
+
+
+test_that("{shinytest2} recording: filter_Asia_male_smoker", {
+  app <- AppDriver$new(name = "filter_Asia_male_smoker", height = 945, width = 1619)
+  app$set_inputs(`tab3_filter_1-study` = c("AUOECAFAOINV", "LIDSAMEVDAISNE", "PVNI", 
+      "SRESSBAVG", "SSEVALCCASCE", "FVNMAMTINP", "ITOVJD", "RDVCUNISPDS", "VEEIIEN", 
+      "VMEATST", "APDLAIA", "ENAH", "FISAGPUSLVF", "PMMERMICI", "STISHV", "ASVA", 
+      "MTOVFMAISV", "PNDEPPIEVEMMISEUFPLUEPNRE", "RILITARLPNELAHSHSTA", "VVSACNACVJDL", 
+      "DPEVMV", "INVNESN", "NAVNMCIFMNNUF", "PARNVF", "UENUFUEVE", "APTIFNID", "ASPAPLMEVDAISNE", 
+      "ESMEMVCOSDSSEDS", "INA", "SAASVEANANESS"))
+  app$set_inputs(`tab3_filter_1-continent` = c("Africa", "Asia", "Europe", "North America", 
+      "Oceania", "South America"))
+  app$set_inputs(`tab5_pac_1-ycate` = "Disease Diagnosis 1")
+  app$set_inputs(`tab5_pac_1-cx2c` = "Alcohol Use Status")
+  app$set_inputs(`tab5_paq_1-qx2c` = "Alcohol Use Status")
+  app$set_inputs(`tab5_paq_1-qx3c` = "Deceased")
+  app$set_inputs(`tab3_filter_1-filtertabs` = "Study Filters")
+  app$set_inputs(`tab4_visual_1-visaultabs` = "Metadata")
+  app$set_inputs(`tab5_pac_1-pactabs` = "Univariate")
+  app$set_inputs(`tab5_paq_1-paqtabs` = "Univariate")
+  app$click("tab2_meta_1-clearmeta")
+  app$click("tab2_meta_1-clearava")
+  app$click("tab3_filter_1-resetallf")
+  app$click("tab3_filter_1-resetsf")
+  app$click("tab3_filter_1-resetpf")
+  app$set_inputs(`tab5_pac_1-cx1q` = "Scale 2")
+  app$set_inputs(`tab5_paq_1-yquan` = "Scale 4")
+  app$set_inputs(`tab5_paq_1-qx1q` = "Age at Assessment")
+  app$set_inputs(`tab3_filter_1-minage` = c(0, 110))
+  app$set_inputs(`tab3_filter_1-studysize` = c(500, 5000))
+  app$set_inputs(`tab3_filter_1-ageatass` = c(0, 110))
+  app$set_inputs(`tab3_filter_1-yod` = c(1990, 2030))
+  app$set_inputs(`tab3_filter_1-bmi` = c(10, 40))
+  app$set_inputs(`tab3_filter_1-scale1` = c(20, 40))
+  app$set_inputs(`tab3_filter_1-scale2` = c(10, 25))
+  app$set_inputs(`tab3_filter_1-scale3` = c(0, 25))
+  app$set_inputs(`tab3_filter_1-scale4` = c(1, 100))
+  app$set_inputs(`tab3_filter_1-access` = c("No", "Yes"))
+  app$set_inputs(`tab3_filter_1-studyfollow` = c("No", "Yes"))
[…]
