### 2024-06-14 #6693 @RoryChenXY
Dear All, @elimillera, @pydemull, @ZekeMarshall  and @crvernon,

I am back. Sorry this is taking forever!

I have now converted the app into a package with {golem}, and created a branch (golem) for the updates.
I hope you can see the content in my branch - I have not updated the main yet.

As my original app name contains an underscore, it is not permitted as an R package name so I created 'datarepexp' as the package name.

By making this app into a package, I have learned a lot.
Dependency: Used {renv} for dependencies, and also DESCRIPTION and NAMESPACE.
Testing: Used {testthat} and {shinytest2}.

I have improved some of my functions and designs, the current version is deployed here:
https://rorychenxy.shinyapps.io/repexp/

I am still working on the documentation but would like to share the progress.
Thank you for your time and patience.

Best,
Rory
