+++ datarepexp/README.Rmd
@@ -20,38 +20,30 @@ knitr::opts_chunk$set(
 [![Codecov test coverage](https://codecov.io/gh/RoryChenXY/DataRepExp_public/branch/main/graph/badge.svg)](https://app.codecov.io/gh/RoryChenXY/DataRepExp_public?branch=main)
 <!-- badges: end -->
 
-The goal of datarepexp is to ...
+[datarepexp](https://github.com/RoryChenXY/DataRepExp_public/datarepexp), is an open-source R Shiny application and an R package built using the [golem](https://github.com/ThinkR-open/golem) framework.
+
+Data Repository Explorer is developed to improve the findability, accessibility, interoperability, and reusability (FAIR) of research data held in a data repository.
+
+The application displays standardised metadata across multiple studies including data availability by categories (such as demographics, medical history, imaging data and genomic data) to allow high-level comparison. It enables users to explore and run preliminary analysis from participants that match certain criteria. In addition, it provides features to export reports and aggregated results for data access application purposes. The application was initially developed for a discipline-specific data-sharing platform, the Dementias Platform Australia (DPAU). Envisioning this work could be utilized by other data repositories in diverse disciplines, this demo application was created using simulated health-related data for demonstration purposes.
+
+It can be modified and utilized by other data repositories by adopting the discipline-specific metadata schema and common variables.
 
 ## Installation
 
 You can install the development version of datarepexp from [GitHub](https://github.com/) with:
 
-``` r
+```{r, eval=FALSE}
 # install.packages("devtools")
-devtools::install_github("RoryChenXY/DataRepExp_public/datarepexp")
+devtools::install_github("RoryChenXY/DataRepExp_public/", subdir="datarepexp")
 ```
 
-## Example
+## Demo
 
-This is a basic example which shows you how to solve a common problem:
+The demo app is deployed on shinyapps.io: https://rorychenxy.shinyapps.io/repexp/
 
-```{r example}
+To run the dev version of this app locally:
+```{r, eval=FALSE}
 library(datarepexp)
-## basic example code
+datarepexp::repexp_app()
 ```
 
-What is special about using `README.Rmd` instead of just `README.md`? You can include R chunks like so:
-
-```{r cars}
-summary(cars)
-```
-
-You'll still need to render `README.Rmd` regularly, to keep `README.md` up-to-date. `devtools::build_readme()` is handy for this.
-
-You can also embed plots, for example:
-
-```{r pressure, echo = FALSE}
-plot(pressure)
-```
-
-In that case, don't forget to commit and push the resulting figure files, so they display on GitHub and CRAN.
