+++ datarepexp/NAMESPACE
@@ -7,7 +7,6 @@ export(fct_inputs)
 export(golem_add_external_resources)
 export(mod_tab1_statement_app)
 export(mod_tab2_meta_app)
-export(mod_tab2_meta_server)
 export(mod_tab3_filter_app)
 export(mod_tab4_visual_app)
 export(mod_tab5_pac_app)
