+++ README.md
@@ -6,9 +6,9 @@ The application displays standardised metadata across multiple studies including
 
 This demo application was created using simulated health-related data for demonstration purposes. It can be modified and utilized by other data repositories by adopting the discipline-specific metadata schema and common variables.
 
--   Source Code: <https://github.com/RoryChenXY/DataRepExp_public>
+-   Source Code: <https://github.com/RoryChenXY/DataRepExp_public/datarepexp>
 -   Web Application: <https://rorychenxy.shinyapps.io/DataRepExp/>
--   Contact: [xinyue.chen1\@unsw.edu.au](mailto:%20xinyue.chen1@unsw.edu.au)
+-   Contact: [xinyue.chen1\@unsw.edu.au](mailto:%20xinyue.chen1@unsw.edu.au) / [rory.chen.xy\@gmail.com](mailto:%20rory.chen.xy@gmail.com)
 -   Paper: [paper.md](/paper/paper.md)
 
 ## How to navigate the application
@@ -39,12 +39,12 @@ All plots can be downloaded. They have interactive features such as zoom, select
 
 ## Deployment Notes
 
-To make modification to the applications for your purposes, R programming skill is required.
+To make modifications to the applications for your purposes, R and Shiny programming skill is required.
 1. The application requires both study-level metadata, and particicipant-level data, and a variable information data frame that contains all variable information. Deatail Notes and scripts used to generate data is included in the repository: [Data Preparation Notes](/notes/1_dataprep.md). 
 2. DataRepExp was built in Shiny modules. Modularity makes the app easy to test, maintain, and deploy. The features can be easily further expanded with loose coupling module design: [Application Development Notes](/notes/2_app_dev_notes.md).
 3. I have created a copy of this application that was modified to add two more variables just as a demonstration to set up your own variables: [How to add new variables](/notes/3_add_new_var.md).
 4. Considering some repositories may hold highly sensitive data, or individual-level data may not be not available, a metadata-only version DataRepExp has also been developed, and relevant code is included [here](/archive/meta_demo).
-5. The Data Repository Explorer, DataRepExp, is hosted through easy-to-use [shinyapps.io](https://www.shinyapps.io/),  while the DPAU version is hosted on AWS environment using Shiny Server for high availability, scalability, security, and compliance.  The detailed deploy instruction can be found [here](https://shiny.posit.co/r/deploy.html)
+5. The Data Repository Explorer, DataRepExp, is hosted through easy-to-use [shinyapps.io](https://www.shinyapps.io/),  while the DPAU version is hosted on AWS environment using Shiny Server for high availability, scalability, security, and compliance.  The detailed deployment instruction can be found [here](https://shiny.posit.co/r/deploy.html)
 
 ## Acknowledgements
 This application was inspired by the visualization tool developed by Dementias Platform UK([DPUK](https://www.dementiasplatform.uk/)) using PowerBI, then developed for the Dementias Platform Australia ([DPAU](https://www.dementiasplatform.com.au/)) in R-Shiny. We acknowledge the generous sharing of best practices and knowledge from DPUK.
