### 2024-07-19 db3cade
Update mod_tab2_meta.R
Adjust the numeric filters with slide to vertical and adjust the position
