### 2024-07-19 fedb167
Update mod_tab2_meta.R
update DT options so the csv files exported have a name that describe more specifically their content.
