### 2024-07-19 2bbfcc7
Update mod_tab4_visual.R
add line breaker so that
metadata that appear on the bar plots can all be viewed in the box when we have the cursor on the graphic.
