+++ datarepexp/NAMESPACE
@@ -30,12 +30,15 @@ importFrom(DT,renderDT)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(collapse,fmax)
 importFrom(collapse,fmin)
+importFrom(desc,desc_get_version)
 importFrom(dplyr,arrange)
 importFrom(dplyr,count)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,mutate)
+importFrom(dplyr,mutate_if)
 importFrom(dplyr,n)
+importFrom(dplyr,relocate)
 importFrom(dplyr,select)
 importFrom(dplyr,summarise)
 importFrom(dplyr,ungroup)
@@ -72,6 +75,8 @@ importFrom(scales,breaks_pretty)
 importFrom(scales,percent)
 importFrom(shiny,NS)
 importFrom(shiny,checkboxGroupInput)
+importFrom(shiny,htmlOutput)
+importFrom(shiny,renderUI)
 importFrom(shiny,shinyApp)
 importFrom(shiny,tagList)
 importFrom(shiny,wellPanel)
