+++ datarepexp/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: datarepexp
 Title: Data Repository Explorer
-Version: 0.0.0.9000
+Version: 0.0.1
 Authors@R: person('Rory', 'Chen', email = 'rory.chen.xy@gmail.com', role = c('cre', 'aut'), comment = c(ORCID = "0000-0002-2094-2682"))
 Description: The Data Repository Explorer, is an open-source R Shiny application developed to improve the 
                      findability, accessibility, interoperability, and reusability (FAIR) of research data held in a data repository.
@@ -21,15 +21,16 @@ Imports:
     ggplot2 (>= 3.5.1),
     plotly (>= 4.10.4),
     scales (>= 1.3.0),
-    collapse (>= 2.0.14),
+    collapse (>= 2.0.15),
     forcats (>= 1.0.0),
-    grDevices (>= 4.4.0),
-    RColorBrewer (>= 1.1.3)
+    grDevices (>= 4.4.1),
+    RColorBrewer (>= 1.1.3),
+    desc (>= 1.4.3)
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 Suggests: 
-    testthat (>= 3.2.1.1),
+    testthat,
     knitr,
     rmarkdown,
     devtools,
