+++ datarepexp/app.R
@@ -1,11 +1,12 @@
 # Launch the ShinyApp (Do not remove this comment)
 # To deploy, run: rsconnect::deployApp()
 # Or use the blue button on top of this file
+pkgload::load_all(export_all = FALSE,
+                  helpers = FALSE,
+                  attach_testthat = FALSE)
 
-devtools::load_all()
-
-options("golem.app.prod" = TRUE)
-
-library(datarepexp)
+options("golem.app.prod" = FALSE)
 
 datarepexp::repexp_app() # add parameters here (if any)
+
+# shinytest2::record_test() #Initial the test
