### 2024-07-23 314c67e
Update golem-config.yml
version
