+++ datarepexp/inst/golem-config.yml
@@ -1,6 +1,6 @@
 default:
   golem_name: datarepexp
-  golem_version: 0.0.0.9000
+  golem_version: 0.0.1
   app_prod: no
 production:
   app_prod: yes
