### 2024-07-26 caa0e9a
Update mod_tab1_statement.R
getting pkg version is not working properly
