### 2024-07-26 0640e96
Update mod_tab5_paq.R
importFrom stats IQR lm quantile sd
