+++ datarepexp/NAMESPACE
@@ -30,7 +30,6 @@ importFrom(DT,renderDT)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(collapse,fmax)
 importFrom(collapse,fmin)
-importFrom(desc,desc_get_version)
 importFrom(dplyr,arrange)
 importFrom(dplyr,count)
 importFrom(dplyr,filter)
@@ -83,6 +82,10 @@ importFrom(shiny,wellPanel)
 importFrom(shinyWidgets,pickerInput)
 importFrom(shinyjs,reset)
 importFrom(shinyjs,useShinyjs)
+importFrom(stats,IQR)
+importFrom(stats,lm)
+importFrom(stats,quantile)
+importFrom(stats,sd)
 importFrom(tidyr,complete)
 importFrom(tidyr,drop_na)
 importFrom(tidyr,pivot_wider)
