### 2024-07-26 4852489
Update fct_visplotly.R
comparing class() to string, use inherits instead
