### 2024-07-26 2d8447a
Update utils_countfactor.R
comparing class() to string, use inherits instead
