### 2024-07-26 a71de5c
Update plotly_histogram_grouped.Rd
format
