+++ datarepexp/man/plotly_histogram_grouped.Rd
@@ -22,7 +22,12 @@ A plotly histogram compare a numeric variable between groups.
 A standardised function to generate a grouped histogram with three display options.
 }
 \examples{
-group_his_df <- data.frame(numvar = rep(1: 20, 5), fctvar = as.factor(c(rep(c("A", "B", "C"), 30), rep("A", 10))))
-plotly_histogram_grouped(df = group_his_df, var = "numvar", groupvar = "fctvar", varlabel = "Numbers", grouplabel = "Groups")
+group_his_df <- data.frame(numvar = rep(1: 20, 5),
+                           fctvar = as.factor(c(rep(c("A", "B", "C"), 30), rep("A", 10))))
+plotly_histogram_grouped(df = group_his_df,
+                         var = "numvar",
+                         groupvar = "fctvar",
+                         varlabel = "Numbers",
+                         grouplabel = "Groups")
 
 }
