+++ datarepexp/DESCRIPTION
@@ -25,7 +25,7 @@ Imports:
     forcats (>= 1.0.0),
     grDevices (>= 4.4.1),
     RColorBrewer (>= 1.1.3),
-    desc (>= 1.4.3)
+    stats (>= 4.4.1)
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.2
