+++ datarepexp/dev/run_dev.R
@@ -1,5 +1,5 @@
 # Set options here
-options(golem.app.prod = FALSE) # TRUE = production mode, FALSE = development mode
+options(golem.app.prod = TRUE) # TRUE = production mode, FALSE = development mode
 
 # Comment this if you don't want the app to be served on a random port
 # options(shiny.port = httpuv::randomPort())
@@ -10,6 +10,8 @@ golem::detach_all_attached()
 
 # Document and reload your package
 golem::document_and_reload()
+
+golem::disable_autoload()
 
 # Run the application
 datarepexp::repexp_app()
