### 2024-07-26 RoryChenXY/DataRepExp_public#2 @RoryChenXY
@elimillera, thank you for your input.
I have utilised {renv} as suggested, as well as the DESCRIPTION file and NAMESPACE for dependency management
