### 2024-08-14 81828fd
Create CONTRIBUTING.md
