+++ paper/paper.bib
@@ -358,4 +358,11 @@ https://github.com/tidyverse/tidyverse},
   doi = {10.21105/joss.01686},
 }
 
+@misc{R-golem,
+  title = {golem: A Framework for Robust Shiny Applications},
+  author = {Colin Fay and Vincent Guyader and Sébastien Rochette and Cervan Girard},
+  year = {2024},
+  note = {R package version 0.5.0, https://github.com/ThinkR-open/golem},
+  url = {https://thinkr-open.github.io/golem/}
+}
 
