### 2024-08-19 2011c2f
Update 02_ppt_simulation.R
library: add magrittr, remove tidyverse
