### 2024-08-19 RoryChenXY/DataRepExp_public#3 @RoryChenXY
The 'simdata.RData' is not standard.

Basically, the app requires 3 data frames: study-level data, participant-level data and a data dictionary.
It still requires some 'hard coding' to get the user data ready and modify the app for their own purposes.

We have two data repository partners who are interested in this app, I will work with them to build a better documentation.
