### 2024-08-19 #6693 @RoryChenXY
Dear All,

I believe I have addressed all the comments and issues to the best of my ability.

I sincerely appreciate the generous contributions of your time and the valuable feedback you’ve provided. My apologies for the delays—academics never seem to have enough time!

This process has been a significant learning experience for me, especially in improving this app and converting it into a package.

Thank you once again for your kind support.

Best regards,
Rory
