### 2024-08-22 #6693 @RoryChenXY
> Hi @RoryChenXY , thanks for addressing the comments above! I have seen that you have pushed these changes to the main branch, would it be possible to make a release? I will then re-visit my checklist :)

Hi @ZekeMarshall, I just made a release - should learn more about these processes. Thanks!
