### 2024-09-09 #6693 @ZekeMarshall
Hi @crvernon and @RoryChenXY , I've just worked through my checklist again using the v0.0.1 release of the app. Thanks for addressing all my comments @RoryChenXY ! Everything is looking great, only two remaining suggestions from me before I can complete my checklist:

1. I have generated the most recent version of the paper and the app screenshot is not showing, could this be fixed?
2. In my last paper I was told to reference R packages using the preferred recommended citation, or using the CRAN DOI. Is the CRAN DOI citation format required @crvernon ?

Cheers!

Zeke
