### 2024-09-09 2ec7a50
Update paper.bib
update R package doi
