+++ paper/paper.bib
@@ -123,7 +123,7 @@
   year = {2023},
   note = {R package version 1.1.4},
   url = {https://dplyr.tidyverse.org},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.dplyr}
 }
 
 @misc{R-DT,
@@ -132,7 +132,7 @@
   year = {2024},
   note = {R package version 0.32},
   url = {https://github.com/rstudio/DT},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.DT}
 }
 
 @misc{R-fontawesome,
@@ -142,7 +142,7 @@
   note = {R package version 0.5.2, 
 https://rstudio.github.io/fontawesome/},
   url = {https://github.com/rstudio/fontawesome},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.fontawesome}
 }
 
 @misc{R-forcats,
@@ -152,7 +152,7 @@ https://rstudio.github.io/fontawesome/},
   note = {R package version 1.0.0, 
 https://github.com/tidyverse/forcats},
   url = {https://forcats.tidyverse.org/},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.forcats}
 }
 
 @book{R-ggplot2,
@@ -172,7 +172,7 @@ https://github.com/tidyverse/ggplot2},
   note = {R package version 0.5.7, 
 https://rstudio.github.io/htmltools/},
   url = {https://github.com/rstudio/htmltools},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.htmltools}
 }
 
 @article{R-lubridate,
@@ -192,7 +192,7 @@ https://github.com/tidyverse/lubridate},
   note = {R package version 2.0.3, 
 https://github.com/tidyverse/magrittr},
   url = {https://magrittr.tidyverse.org},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.magrittr}
 }
 
 @book{R-plotly,
@@ -210,7 +210,7 @@ https://github.com/tidyverse/magrittr},
   year = {2023},
   note = {R package version 1.0.2},
   url = {https://purrr.tidyverse.org/},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.purrr}
 }
 
 @misc{R-readr,
@@ -219,7 +219,7 @@ https://github.com/tidyverse/magrittr},
   year = {2024},
   note = {R package version 2.1.5},
   url = {https://readr.tidyverse.org},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.readr}
 }
 
 @misc{R-scales,
@@ -228,7 +228,7 @@ https://github.com/tidyverse/magrittr},
   year = {2023},
   note = {R package version 1.3.0},
   url = {https://scales.r-lib.org},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.scales}
 }
 
 @misc{R-shiny,
@@ -237,7 +237,7 @@ https://github.com/tidyverse/magrittr},
   year = {2023},
   note = {R package version 1.8.0},
   url = {https://shiny.posit.co/},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.shiny}
 }
 
 @misc{R-shinydashboard,
@@ -246,7 +246,7 @@ https://github.com/tidyverse/magrittr},
   year = {2021},
   note = {R package version 0.7.2},
   url = {http://rstudio.github.io/shinydashboard/},
-  note = {No DOI identified},
+  doi = {10.32614/CRAN.package.shinydashboard}
 }
 
 @misc{R-shinyjs,
[…]
