### 2024-09-09 #6693 @RoryChenXY
Hi @ZekeMarshall, Thank you for your feedback.

1. screenshot - fixed, it was the lowercase/ uppercase error
2. DOI: I update the package DOIs like '10.32614/CRAN.package.golem', didn't realise they were available.

Huge thanks for all!

Best
Rory
