### 2024-09-16 #6693 @RoryChenXY
Hi @crvernon, I hope you are well.

Our centre has a media office, they would like to prepare a media release and have enquired if any update to the publication date could be advised in advance.

Thank you so much!

Best,
Rory
