### 2024-09-18 #6693 @RoryChenXY
Hi @crvernon,

Apologies for the ignorance.
I have now made a tagged release (v1.0.0) and deposited it into Zenodo and hope we are ready to do:

Chen, R. (2024). DataRepExp. In Journal of Open Source Software (v1.0.0). Zenodo. https://doi.org/10.5281/zenodo.13777122

Please let me know if any further information is required.

Best wishes,
Rory
