### 2024-09-27 #6693 @crvernon
:wave: @RoryChenXY - thanks for setting up the archive and version!  Here are a few things I need you to address:

**In the Zenodo archive:**

- [ ] The 1) tile of the archive and 2) the authors and the order in which they appear need to match what is in your paper exactly.  Also, the license you specify in the Zenodo archive should match the one you use in your GitHub repository for the code base.  Could you edit the metadata in the Zenodo archive to fix these three things?  No need to create another release.

**In your paper:**

- [ ] LINES 21-26:  I would cut these lines from the paper...so, ending after your CHeBA citation.  All of this information can be found in your repository which will be available alongside your paper when it gets published.
- [ ] Throughout the paper, you only need to cite the Wilkinson et al. paper after the FAIR acronym on the first occurrence.  The same goes for the CHeBA citation for DPAU.
- [ ] LINES 66-68:  I would remove the statement: "R programming skill is required for reproducibility, detailed documentation and syntax is open-source and publicly available." as it should be assumed by the reader.
- [ ] LINE 88:  Missing period at the end of the line.
- [ ] LINE 103:  Should "integrative" be "interactive" here?
- [ ] Please fix references that have characters that need to be capitalized.  For example, in LINE 121, "r" should be capitalized.  There are others as well so take a look at these and fix them where necessary.  You can maintain formatting of particular characters in your bib file by using curly brackets around the characters you wish to maintain the formatting of.

There is no need to create another release after you make these changes.  Once you have done these, let me know and I will accept this for publication.

Thanks!
