+++ paper/paper.md
@@ -13,22 +13,31 @@ authors:
     orcid: 0000-0002-2094-2682
     affiliation: 1 
   - name: Vibeke S Catts
+    orcid: 0000-0002-9892-0547
     affiliation: 1
   - name: Ashleigh Vella
+    orcid: 0000-0002-5653-5077
     affiliation: 1
   - name: Juan Carlo San Jose
+    orcid: 0000-0002-2365-9153
     affiliation: 2
   - name: Sarah Bauermeister
+    orcid: 0000-0001-9463-6971
     affiliation: "3, 4"
-  - name: Josh Bauermeister
+  - name: Joshua Bauermeister
+    orcid: 0000-0002-9154-1392
     affiliation: "3, 4"
   - name: Emma Squires
+    orcid: 0000-0003-0693-5846
     affiliation: "3, 5"
   - name: Simon Thompson
+    orcid: 0000-0001-6799-0705
     affiliation: "3, 5"
   - name: John Gallacher
+    orcid: 0000-0002-2394-5299
     affiliation: "3, 4"
   - name: Perminder S. Sachdev
+    orcid: 0000-0002-9595-3220
     affiliation: 1  
 affiliations:
  - name: Centre for Healthy Brain Ageing, Discipline of Psychiatry & Mental Health, School of Clinical Medicine, University of New South Wales, Sydney, Australia
@@ -41,7 +50,7 @@ affiliations:
    index: 4
  - name: Population Data Science, Swansea University, Swansea, United Kingdom
    index: 5
-date: 19 August 2024
+date: 30 September 2024
 bibliography: paper.bib
 
 ---
@@ -52,11 +61,6 @@ The application displays standardised metadata across multiple studies including
 It enables users to explore and run preliminary analysis from participants that match certain criteria.
 In addition, it provides features to export reports and aggregated results for data access application purposes.
 The application was initially developed for a discipline-specific data-sharing platform, the Dementias Platform Australia (DPAU) [@DPAU].
-Envisioning this work could be utilized by other data repositories in diverse disciplines, this demo application was created using simulated health-related data for demonstration purposes.
-
--   Source Code: <https://github.com/RoryChenXY/DataRepExp_public>
--   Web Application: <https://rorychenxy.shinyapps.io/DataRepExp/>
--   Contact: [xinyue.chen1\@unsw.edu.au](mailto:%20xinyue.chen1@unsw.edu.au)
 
 
 # Statement of Need
@@ -64,17 +68,17 @@ Data repositories have become increasingly important in recent years as more emp
 By making data publicly available through repositories, researchers can ensure data persistence and support data preservation, as well as facilitate the reuse of their data, thereby increasing the potential for new scientific discoveries.
 
 
-However, challenges exist for data findability, accessibility, interoperability, and reusability (FAIR) [@FAIR].
+However, challenges exist for data findability, accessibility, interoperability, and reusability (FAIR).
 Even though most data repositories have adopted various metadata schemas to describe the dataset [@Contaxis2022], it is increasingly a challenge for researchers to find relevant data that meet research interests or needs [@Gregory2018].
 For multi-study research, applying to access different datasets usually comes with diverse and complicated data-sharing requirements and workflows, extensive administrative workloads and waiting periods.
-Upon approval, substantial efforts of data harmonization are usually required due to inconsistent data structures and labeling conventions, and harmonised dataset are hardly reused.
+Upon approval, substantial efforts of data harmonization are usually required due to inconsistent data structures and labelling conventions, and harmonised datasets are hardly reused.
 We found that many data repositories do not provide comprehensive metadata, nor centralised tables for comparison.
 With repositories that provide data visualisation, Power BI and Tableau are commonly used but costs are incurred. 
 R-shiny could provide more flexibility and functions at a fraction of the cost.
 
 
[…]
