+++ paper/paper.bib
@@ -58,7 +58,7 @@
 }
 
 @article{bauermeister2023ready,
-  title={Research-ready data: the C-Surv data model},
+  title={Research-ready data: the {C-Surv} data model},
   author={Bauermeister, Sarah and Bauermeister, Joshua R and Bridgman, Ruth and Felici, Caterina and Newbury, Mark and North, Laura and Orton, Christopher and Squires, Emma and Thompson, Simon and Young, Simon and others},
   journal={European Journal of Epidemiology},
   volume={38},
@@ -71,9 +71,9 @@
 
 
 @article{bauermeister2020dementias,
-  title={The Dementias Platform UK (DPUK) data portal},
+  title={The {Dementias Platform UK (DPUK)} data portal},
   author={Bauermeister, Sarah and Orton, Christopher and Thompson, Simon and Barker, Roger A and Bauermeister, Joshua R and Ben-Shlomo, Yoav and Brayne, Carol and Burn, David and Campbell, Archie and Calvin, Catherine and others},
-  journal={European journal of epidemiology},
+  journal={European Journal of Epidemiology},
   volume={35},
   pages={601--611},
   year={2020},
@@ -98,7 +98,7 @@
 } 
 
 @misc{R-Base,
-  title = {R: A Language and Environment for Statistical Computing},
+  title = {{R: A Language and Environment for Statistical Computing}},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
@@ -107,18 +107,9 @@
   note = {No DOI identified},
 }
 
-@misc{R-base,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2024},
-  url = {https://www.R-project.org/},
-  note = {No DOI identified},
-}
 
 @misc{R-dplyr,
-  title = {dplyr: A Grammar of Data Manipulation},
+  title = {{dplyr: A Grammar of Data Manipulation}},
   author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller and Davis Vaughan},
   year = {2023},
   note = {R package version 1.1.4},
@@ -127,7 +118,7 @@
 }
 
 @misc{R-DT,
-  title = {DT: A Wrapper of the JavaScript Library DataTables},
+  title = {{DT: A Wrapper of the JavaScript Library DataTables}},
   author = {Yihui Xie and Joe Cheng and Xianying Tan},
   year = {2024},
   note = {R package version 0.32},
@@ -136,7 +127,7 @@
 }
 
 @misc{R-fontawesome,
-  title = {fontawesome: Easily Work with Font Awesome Icons},
+  title = {{fontawesome: Easily Work with Font Awesome Icons}},
   author = {Richard Iannone},
   year = {2023},
   note = {R package version 0.5.2, 
@@ -146,7 +137,7 @@ https://rstudio.github.io/fontawesome/},
 }
 
 @misc{R-forcats,
-  title = {forcats: Tools for Working with Categorical Variables (Factors)},
+  title = {{forcats: Tools for Working with Categorical Variables (Factors)}},
   author = {Hadley Wickham},
   year = {2023},
   note = {R package version 1.0.0, 
@@ -156,7 +147,7 @@ https://github.com/tidyverse/forcats},
 }
 
 @book{R-ggplot2,
-  title = {ggplot2: Elegant Graphics for Data Analysis},
[…]
