### 2024-09-30 #6693 @RoryChenXY
@crvernon Hi Chris! Thank you for the feedback.
I have updated the Zenodo and the paper - except: 'LINE 103: Should "integrative" be "interactive" here?' - It is interactive - meant for the functions enabled with Plotly and DT.

Really appreciate your support!

Cheers,
Rory
