### 2024-10-01 #6693 @crvernon
🥳 Congratulations on your new publication @RoryChenXY! Many thanks to @elimillera, @pydemull and @ZekeMarshall for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
