### 2024-10-01 #6693 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
